Moved Apple Pay controllers
